:truck: Renames use effect on update only hook
