+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
-import { useEffectOnUpdateOnly } from './custom-use-effects/useEffectOnUpdateOnly'
+import { useEffectOnUpdateOnly } from './custom-use-effects/use-effect-on-update-only'
 import { useQuery } from './data-fetching/useQuery'
 import { useFetchAll } from './data-fetching/useFetchAll'
 import { useFetchOne } from './data-fetching/useFetchOne'
